Fragments are common parts of entities. When entities are build from JSON elorm-gen adds fields from fragments and register fragment in special list Fragments per entity type.
Event handlers for any entity def not are list of functions. Developer can define as many handlers as he wants. Handler can be defined on factory level for particular entity def or for all defs that contains a particular fragment. It's useful when your fragments defines parts of some functionality. and it's good for code reusing.